requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.11.17
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
fastapi==0.105.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
idna==3.6
iniconfig==2.0.0
numpy==1.26.2
packaging==23.2
pandas==2.1.4
pluggy==1.3.0
pyasn1==0.5.1
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.14.5
pymongo==4.6.1
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
pythondi==1.2.4
pytz==2023.3.post1
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.9.0
tzdata==2023.3
uvicorn==0.24.0.post1
watchfiles==0.21.0
websockets==12.0
